requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
anthropic==0.39.0
pymupdf==1.24.10
python-pptx==1.0.2
python-multipart==0.0.9
python-dotenv==1.0.1
openpyxl==3.1.5
jinja2==3.1.4
